# src/main/resources/static/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Next Gen Symmetric Apps",
      "version": "1.0"
    },
    "paths": {
...
      "{{url-prefix}}/api/actors": {
        "get": {
          "tags": [ "Actors" ],
          "summary": "Search Actors",
          "description": "Returns an array of `Actor` documents",
          "parameters": [
            {
              "name": "q",
              "in": "query",
              "description": "Filter by search on Actor.Name<br><br>`valid: 2 - 20 characters`",
              "required": false,
              "style": "form",
              "explode": true,
              "schema": {
                "maxLength": 20,
                "minLength": 2,
                "type": "string",
                "example": "reeves"
              }
            },
            {
              "name": "pageNumber",
              "in": "query",
              "description": "1 based page number<br><br>`valid: 1 - 10,000`",
              "required": false,
              "style": "form",
              "explode": true,
              "schema": {
                "maximum": 10000,
                "minimum": 1,
                "example": 1,
                "type": "integer",
                "default": 1
              }
            },
            {
              "name": "pageSize",
              "in": "query",
              "description": "Limit to pageSize documents<br><br>`valid 1 - 1000`",
              "required": false,
              "style": "form",
              "explode": true,
              "schema": {
                "maximum": 1000,
                "minimum": 1,
                "example": 10,
                "type": "integer",
                "default": 100
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Array of `Actor` documents<br>Empty array if not found",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Actor"
                    }
                  },
                  "example": [
                    {
                      "id": "nm0000206",
                      "actorId": "nm0000206",
                      "partitionKey": "6",
                      "type": "Actor",
                      "name": "Keanu Reeves",
                      "birthYear": 1964,
                      "textSearch": "keanu reeves",
                      "profession": [
                        "actor",
                        "producer",
                        "soundtrack"
                      ],
                      "movies": [
                        {
                          "movieId": "tt0102685",
                          "title": "Point Break",
                          "year": 1991,
                          "runtime": 122,
                          "genres": [
                            "Action",
                            "Crime",
                            "Thriller"
                          ]
                        },
                        {
                          "movieId": "tt0133093",
                          "title": "The Matrix",
                          "year": 1999,
                          "runtime": 136,
                          "genres": [
                            "Action",
                            "Sci-Fi"
                          ]
                        },
                        {
                          "movieId": "tt0234215",
                          "title": "The Matrix Reloaded",
                          "year": 2003,
                          "runtime": 138,
                          "genres": [
                            "Action",
                            "Sci-Fi"
                          ]
                        },
                        {
                          "movieId": "tt0242653",
                          "title": "The Matrix Revolutions",
                          "year": 2003,
                          "runtime": 129,
                          "genres": [
                            "Action",
                            "Sci-Fi"
                          ]
                        }
                      ]
                    },
                    {
                      "id": "nm0716257",
                      "actorId": "nm0716257",
                      "partitionKey": "7",
                      "type": "Actor",
                      "name": "Matt Reeves",
                      "birthYear": 1966,
                      "textSearch": "matt reeves",
                      "profession": [
                        "producer",
                        "writer",
                        "director"
                      ],
                      "movies": [
                        {
                          "movieId": "tt3450958",
                          "title": "War for the Planet of the Apes",
                          "year": 2017,
                          "runtime": 140,
                          "genres": [
                            "Action",
                            "Adventure",
                            "Drama"
                          ]
                        }
                      ]
                    },
                    {
                      "id": "nm0716293",
                      "actorId": "nm0716293",
                      "partitionKey": "3",
                      "type": "Actor",
                      "name": "Saskia Reeves",
                      "birthYear": 1961,
                      "textSearch": "saskia reeves",
                      "profession": [
                        "actress",
                        "soundtrack"
                      ],
                      "movies": [
                        {
                          "movieId": "tt0101595",
                          "title": "Close My Eyes",
                          "year": 1991,
                          "runtime": 108,
                          "genres": [
                            "Drama",
                            "Romance"
                          ]
                        }
                      ]
                    }
                  ]
                }
              }
            },
            "400": {
              "description": "Invalid parameter",
              "content": {
                "application/json": {
                  "example": {
                    "type": "https://github.com/retaildevcrews/ngsa/blob/main/docs/ParameterValidation.md#actors-api",
                    "title": "Parameter validation error",
                    "detail": "One or more invalid parameters were specified.",
                    "status": 400,
                    "instance": "/api/actors?q=#%&pageNumber=1&pageSize=32033",
                    "validationErrors": [
                      {
                        "code": "InvalidValue",
                        "target": "pageSize",
                        "message": "The parameter 'pageSize' should be between 1 and 1000."
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "{{url-prefix}}/api/actors/{actorId}": {
        "get": {
          "tags": [ "Actors" ],
          "summary": "Get Actor by ID",
          "description": "Returns an `Actor` document by actorId",
          "parameters": [
            {
              "name": "actorId",
              "in": "path",
              "description": "`valid: begins with nm; followed by 5 - 7 digits`<br><br>example: Keanu Reeves",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "string"
              },
              "example": "nm0000206"
            }
          ],
          "responses": {
            "200": {
              "description": "`Actor`",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Actor"
                  },
                  "example": {
                    "id": "nm0000206",
                    "actorId": "nm0000206",
                    "type": "Actor",
                    "name": "Keanu Reeves",
                    "birthYear": 1964,
                    "profession": [ "actor", "producer", "soundtrack" ],
                    "movies": [
                      {
                        "movieId": "tt0102685",
                        "title": "Point Break",
                        "year": 1991,
                        "runtime": 122,
                        "genres": [ "Action", "Crime", "Thriller" ]
                      },
                      {
                        "movieId": "tt0133093",
                        "title": "The Matrix",
                        "year": 1999,
                        "runtime": 136,
                        "genres": [ "Action", "Sci-Fi" ]
                      },
                      {
                        "movieId": "tt0234215",
                        "title": "The Matrix Reloaded",
                        "year": 2003,
                        "runtime": 138,
                        "genres": [ "Action", "Sci-Fi" ]
                      },
                      {
                        "movieId": "tt0242653",
                        "title": "The Matrix Revolutions",
                        "year": 2003,
                        "runtime": 129,
                        "genres": [ "Action", "Sci-Fi" ]
                      },
                      {
                        "movieId": "tt2911666",
                        "title": "John Wick",
                        "year": 2014,
                        "runtime": 101,
                        "genres": [ "Action", "Crime", "Thriller" ]
                      },
                      {
                        "movieId": "tt4425200",
                        "title": "John Wick: Chapter 2",
                        "year": 2017,
                        "runtime": 122,
                        "genres": [ "Action", "Crime", "Thriller" ]
                      },
                      {
                        "movieId": "tt6146586",
                        "title": "John Wick: Chapter 3 - Parabellum",
                        "year": 2019,
                        "runtime": 131,
                        "genres": [ "Action", "Crime", "Thriller" ]
                      }
                    ]
                  }
                }
              }
            },
            "400": {
              "description": "Invalid actorId",
              "content": {
                "application/json": {
                  "example": {
                    "type": "https://github.com/retaildevcrews/ngsa/blob/main/docs/ParameterValidation.md#actors-direct-read",
                    "title": "Parameter validation error",
                    "detail": "One or more invalid parameters were specified.",
                    "status": 400,
                    "instance": "/api/actors/123",
                    "validationErrors": [
                      {
                        "code": "InvalidValue",
                        "target": "actorId",
                        "message": "The parameter 'actorId' should start with 'nm' and be between 7 and 11 characters in total."
                      }
                    ]
                  }
                }
              }
            },
            "404": {
              "description": "actorId not found",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "string",
                    "example": "actorId not found"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "Actor": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "actorId": {
              "type": "string"
            },
            "partitionKey": {
              "type": "string"
            },
            "type": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "birthYear": {
              "type": "integer"
            },
            "deathYear": {
              "type": "integer",
              "nullable": true
            },
            "textSearch": {
              "type": "string"
            },
            "profession": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "movies": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "movieId": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "year": {
                    "type": "integer"
                  },
                  "runtime": {
                    "type": "integer"
                  },
                  "genres": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "additionalProperties": false
        },
...
      }
    }
  }